Update sales-related tests
